+++ realm/realm-library/src/main/java/io/realm/Realm.java
@@ -1515,7 +1515,6 @@ public class Realm extends BaseRealm {
      * called automatically. If any exception is thrown during the transaction {@link #cancelTransaction()} will be
      * called instead of {@link #commitTransaction()}.
      * <p>
-     *
      * Calling this method from the UI thread will throw a {@link RealmException}. Doing so may result in a drop of frames
      * or even ANRs. We recommend calling this method from a non-UI thread or using
      * {@link #executeTransactionAsync(Transaction)} instead.
@@ -1523,7 +1522,7 @@ public class Realm extends BaseRealm {
      * @param transaction the {@link io.realm.Realm.Transaction} to execute.
      * @throws IllegalArgumentException if the {@code transaction} is {@code null}.
      * @throws RealmMigrationNeededException if the latest version contains incompatible schema changes.
-     * @throws RealmException if called from the UI thread, unless an explicit opt-in has been declared in the configuration.
+     * @throws RealmException if called from the UI thread, unless an explicit opt-in has been declared in {@link RealmConfiguration.Builder#allowWritesOnUiThread(boolean)}.
      */
     public void executeTransaction(Transaction transaction) {
         //noinspection ConstantConditions