+++ realm/realm-library/src/main/java/io/realm/RealmQuery.java
@@ -53,12 +53,11 @@ import io.realm.internal.fields.FieldDescriptor;
  * <p>
  * A RealmQuery cannot be passed between different threads.
  * <p>
- * Queries can be launched from any thread and results are obtained fairly quickly most of the times. However, launching
- * heavy queries from the UI thread may result in a drop of frames or even ANRs. If you want to prevent these behaviors,
- * you can instantiate a Realm using a {@link RealmConfiguration} that explicitly sets
- * {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)} to {@code false}. This way you will be forced to
- * launch your queries from a non-UI thread. Alternatively, you can also use {@link #findAllAsync()} or
- * {@link #findFirstAsync()}.
+ * Results are obtained fairly quickly most of the times. However, launching heavy queries from the UI thread may result
+ * in a drop of frames or even ANRs. If you want to prevent these behaviors, you can instantiate a Realm using a
+ * {@link RealmConfiguration} that explicitly sets {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)} to
+ * {@code false}. This way queries will be forced to be launched from a non-UI thread. Alternatively, you can also use
+ * {@link #findAllAsync()} or {@link #findFirstAsync()}.
  *
  * @param <E> the class of the objects to be queried.
  * @see <a href="http://en.wikipedia.org/wiki/Builder_pattern">Builder pattern</a>
@@ -1833,9 +1832,11 @@ public class RealmQuery<E> {
      * for the given field, {@code 0} will be returned. When computing the sum, objects with {@code null} values
      * are ignored.
      * @throws java.lang.IllegalArgumentException if the field is not a number type.
+     * @throws RealmException if called from the UI thread after opting out via {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)}.
      */
     public Number sum(String fieldName) {
         realm.checkIfValid();
+        realm.checkAllowQueriesOnUiThread();
 
         long columnKey = schema.getAndCheckFieldColumnKey(fieldName);
         switch (table.getColumnType(columnKey)) {
@@ -1862,9 +1863,11 @@ public class RealmQuery<E> {
      * types of number fields. If no objects exist or they all have {@code null} as the value for the given field,
      * {@code 0} will be returned. When computing the average, objects with {@code null} values are ignored.
      * @throws java.lang.IllegalArgumentException if the field is not a number type.
+     * @throws RealmException if called from the UI thread after opting out via {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)}.
      */
     public double average(String fieldName) {
         realm.checkIfValid();
+        realm.checkAllowQueriesOnUiThread();
 
         long columnIndex = schema.getAndCheckFieldColumnKey(fieldName);
         switch (table.getColumnType(columnIndex)) {
@@ -1887,9 +1890,11 @@ public class RealmQuery<E> {
      * @return the average for the given field amongst objects in query results. This will be of type Decimal128. If no objects exist or they all have {@code null}
      * as the value for the given field {@code 0} will be returned. When computing the average, objects with {@code null} values are ignored.
      * @throws java.lang.IllegalArgumentException if the field is not a Decimal128 type.
+     * @throws RealmException if called from the UI thread after opting out via {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)}.
      */
     public @Nullable Decimal128 averageDecimal128(String fieldName) {
         realm.checkIfValid();
+        realm.checkAllowQueriesOnUiThread();
 
         long columnIndex = schema.getAndCheckFieldColumnKey(fieldName);
         return query.averageDecimal128(columnIndex);
@@ -1902,10 +1907,12 @@ public class RealmQuery<E> {
      * returned. Otherwise the minimum value is returned. When determining the minimum value, objects with {@code null}
      * values are ignored.
      * @throws java.lang.IllegalArgumentException if the field is not a number type.
+     * @throws RealmException if called from the UI thread after opting out via {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)}.
      */
     @Nullable
     public Number min(String fieldName) {
         realm.checkIfValid();
+        realm.checkAllowQueriesOnUiThread();
 
         long columnIndex = schema.getAndCheckFieldColumnKey(fieldName);
         switch (table.getColumnType(columnIndex)) {
@@ -1931,10 +1938,12 @@ public class RealmQuery<E> {
      * will be returned. Otherwise the minimum date is returned. When determining the minimum date, objects with
      * {@code null} values are ignored.
      * @throws java.lang.UnsupportedOperationException if the query is not valid ("syntax error").
+     * @throws RealmException if called from the UI thread after opting out via {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)}.
      */
     @Nullable
     public Date minimumDate(String fieldName) {
         realm.checkIfValid();
+        realm.checkAllowQueriesOnUiThread();
 
         long columnIndex = schema.getAndCheckFieldColumnKey(fieldName);
         return this.query.minimumDate(columnIndex);
@@ -1948,10 +1957,12 @@ public class RealmQuery<E> {
      * returned. Otherwise the maximum value is returned. When determining the maximum value, objects with {@code null}
      * values are ignored.
      * @throws java.lang.IllegalArgumentException if the field is not a number type.
+     * @throws RealmException if called from the UI thread after opting out via {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)}.
      */
     @Nullable
     public Number max(String fieldName) {
         realm.checkIfValid();
+        realm.checkAllowQueriesOnUiThread();
 
         long columnIndex = schema.getAndCheckFieldColumnKey(fieldName);
         switch (table.getColumnType(columnIndex)) {
@@ -1981,6 +1992,7 @@ public class RealmQuery<E> {
     @Nullable
     public Date maximumDate(String fieldName) {
         realm.checkIfValid();
+        realm.checkAllowQueriesOnUiThread();
 
         long columnIndex = schema.getAndCheckFieldColumnKey(fieldName);
         return this.query.maximumDate(columnIndex);
@@ -1991,9 +2003,12 @@ public class RealmQuery<E> {
      *
      * @return the number of matching objects.
      * @throws java.lang.UnsupportedOperationException if the query is not valid ("syntax error").
+     * @throws RealmException if called from the UI thread after opting out via {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)}.
      */
     public long count() {
         realm.checkIfValid();
+        realm.checkAllowQueriesOnUiThread();
+
         // The fastest way of doing `count()` is going through `TableQuery.count()`. Unfortunately
         // doing this does not correctly apply all side effects of queries (like subscriptions). Also
         // some queries constructs, like doing distinct is not easily supported this way.
@@ -2014,14 +2029,13 @@ public class RealmQuery<E> {
      *
      * @return a {@link io.realm.RealmResults} containing objects. If no objects match the condition, a list with zero
      * objects is returned.
-     * @throws RealmException if a query is launched from the UI thread after having explicitly set
-     * {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)} to {@code false}.
+     * @throws RealmException if called from the UI thread after opting out via {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)}.
      * @see io.realm.RealmResults
      */
     @SuppressWarnings("unchecked")
     public RealmResults<E> findAll() {
-        realm.checkAllowQueriesOnUiThread();
         realm.checkIfValid();
+        realm.checkAllowQueriesOnUiThread();
         return createRealmResults(query, queryDescriptors, true);
     }
 
@@ -2258,8 +2272,7 @@ public class RealmQuery<E> {
      * method will throw a {@link RealmException}. Alternatively, you can use {@link #findFirstAsync()}.
      *
      * @return the object found or {@code null} if no object matches the query conditions.
-     * @throws RealmException if a query is launched from the UI thread after having explicitly set
-     * {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)} to {@code false}.
+     * @throws RealmException if called from the UI thread after opting out via {@link RealmConfiguration.Builder#allowQueriesOnUiThread(boolean)}.
      * @see io.realm.RealmObject
      */
     @Nullable